+++ module/src/main/java/com/annimon/tgbotsmodule/api/methods/stickers/AddAnimatedStickerToSetMethod.java
@@ -28,12 +28,12 @@ public class AddAnimatedStickerToSetMethod implements
     }
 
     @Override
-    public Integer getUserId() {
+    public Long getUserId() {
         return method.getUserId();
     }
 
     @Override
-    public AddAnimatedStickerToSetMethod setUserId(@NotNull Integer userId) {
+    public AddAnimatedStickerToSetMethod setUserId(@NotNull Long userId) {
         method.setUserId(userId);
         return this;
     }
